implemented UserIDAPIDatastore
