+++ TwitterClientApp/UserAPIDatastore.swift
@@ -13,9 +13,6 @@ import RxSwift
 protocol UserAPIDatastoreType {
     func getFollowers(userID: Int?, screenName: String?, cursor: Int?, requestNumberOfFollwers: Int?, skipStatus: Bool?, includeEntities: Bool?)
         -> Observable<[String: Any]>
-    
-    func getFollowersID(userID: Int?, screenName: String?, cursor: Int?, stringifyIDs: Bool?, requestNumberOfFollwers: Int?)
-        -> Observable<[String: Any]>
 }
 
 struct UserAPIDatastore: UserAPIDatastoreType {
@@ -24,13 +21,7 @@ struct UserAPIDatastore: UserAPIDatastoreType {
             .GetFollowers(userID: userID, screenName: screenName, cursor: cursor, requestNumberOfFollwers: requestNumberOfFollwers, skipStatus: skipStatus, includeEntities: includeEntities)
             .sendRequest()
     }
-    
-    func getFollowersID(userID: Int? = nil, screenName: String?, cursor: Int?, stringifyIDs: Bool? = false, requestNumberOfFollwers: Int? = 5000)
-        ->Observable<[String: Any]> { return UserRequest
-            .GetFollowersID(userID: userID, screenName: screenName, cursor: cursor, stringifyIDs: stringifyIDs, requestNumberOfFollwers: requestNumberOfFollwers)
-            .sendRequest()
-        }
-    }
+}
 
 fileprivate struct UserRequest {
       fileprivate struct GetFollowers: RequestType {
@@ -70,43 +61,6 @@ fileprivate struct UserRequest {
             
             if let includeEntities = includeEntities {
                 params["include_user_entities"] = includeEntities
-            }
-            return params
-        }
-    }
-    
-      fileprivate struct GetFollowersID: RequestType {
-        typealias Response = [String: Any]
-        
-        fileprivate let userID: Int?
-        fileprivate let screenName: String?
-        fileprivate let cursor: Int?
-        fileprivate let stringifyIDs: Bool?
-        fileprivate let requestNumberOfFollwers: Int?
-        
-        var path: String { return "followers/ids.json" }
-        var method: OAuthSwiftHTTPRequest.Method { return .GET }
-        var parameters: OAuthSwift.Parameters? {
-            var params: [String: Any] = [:]
-            
-            if let userID = userID {
-                params["userID"] = userID
-            }
-            
-            if let screenName = screenName {
-                params["screen_name"] = screenName
-            }
-            
-            if let cursor = cursor {
-                params["cursor"] = cursor
-            }
-            
-            if let requestNumberOfFollwers = requestNumberOfFollwers {
-                params["count"] = requestNumberOfFollwers
-            }
-            
-            if let stringifyIDs = stringifyIDs {
-                params["stringify_ids"] = stringifyIDs
             }
             return params
         }
